move functional DocumentManager tests
