--- src/Sulu/Component/DocumentManager/Tests/symfony-di/core.xml
@@ -1,86 +1,0 @@
-<?xml version="1.0" ?>
-
-<container xmlns="http://symfony.com/schema/dic/services"
-        xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-        xsi:schemaLocation="http://symfony.com/schema/dic/services http://symfony.com/schema/dic/services/services-1.0.xsd">
-
-    <services>
-
-        <!-- Core -->
-        <service id="sulu_document_manager.event_dispatcher.debug" class="Sulu\Component\DocumentManager\EventDispatcher\DebugEventDispatcher" public="false">
-            <argument type="service" id="debug.stopwatch"/>
-            <argument type="service" id="logger"/>
-
-            <tag name="monolog.logger" channel="sulu_document_manager"/>
-        </service>
-
-        <service id="sulu_document_manager.event_dispatcher.standard" class="Symfony\Component\EventDispatcher\EventDispatcher" public="false"/>
-
-        <service id="sulu_document_manager.document_manager" class="Sulu\Component\DocumentManager\DocumentManager" public="true">
-            <argument type="service" id="sulu_document_manager.event_dispatcher"/>
-            <argument type="service" id="sulu_document_manager.node_manager"/>
-        </service>
-
-        <service id="sulu_document_manager.document_registry" class="Sulu\Component\DocumentManager\DocumentRegistry" public="false">
-            <argument>%sulu_document_manager.default_locale%</argument>
-        </service>
-
-        <service id="sulu_document_manager.node_manager" class="Sulu\Component\DocumentManager\NodeManager" public="false">
-            <argument type="service" id="doctrine_phpcr.session"/>
-        </service>
-
-        <service id="sulu_document_manager.node_helper" class="Sulu\Component\DocumentManager\NodeHelper"/>
-
-        <service id="sulu_document_manager.metadata_factory.base" class="Sulu\Component\DocumentManager\Metadata\BaseMetadataFactory" public="true">
-            <argument type="service" id="sulu_document_manager.event_dispatcher" />
-            <argument>%sulu_document_manager.mapping%</argument>
-        </service>
-
-        <service id="sulu_document_manager.metadata_factory" class="Sulu\Component\DocumentManager\Metadata\MetadataFactory" public="false">
-            <argument type="service" id="sulu_document_manager.metadata_factory.base"/>
-        </service>
-
-        <service id="sulu_document_manager.slugifier" class="Symfony\Cmf\Api\Slugifier\CallbackSlugifier" public="false">
-            <argument>Ferrandini\Urlizer::urlize</argument>
-        </service>
-
-        <service id="sulu_document_manager.node_name_slugifier"
-                 class="Sulu\Component\DocumentManager\Slugifier\NodeNameSlugifier" public="false">
-            <argument type="service" id="sulu_document_manager.slugifier"/>
-        </service>
-
-        <service id="sulu_document_manager.namespace_registry" class="Sulu\Component\DocumentManager\NamespaceRegistry" public="false">
-            <argument>%sulu_document_manager.namespace_mapping%</argument>
-        </service>
-
-        <service id="sulu_document_manager.property_encoder" class="Sulu\Component\DocumentManager\PropertyEncoder" public="false">
-            <argument type="service" id="sulu_document_manager.namespace_registry"/>
-        </service>
-
-        <service id="sulu_document_manager.name_resolver" class="Sulu\Component\DocumentManager\NameResolver" public="false"/>
-
-        <!-- Utilities -->
-        <service id="sulu_document_manager.path_segment_registry" class="Sulu\Component\DocumentManager\PathSegmentRegistry">
-            <argument type="collection">
-            </argument>
-        </service>
-
-        <service id="sulu_document_manager.document_inspector" class="Sulu\Component\DocumentManager\DocumentInspector" public="true">
-            <argument type="service" id="sulu_document_manager.document_registry"/>
-            <argument type="service" id="sulu_document_manager.path_segment_registry"/>
-            <argument type="service" id="sulu_document_manager.proxy_factory"/>
-        </service>
-
-        <!-- Proxy manager -->
-        <service id="sulu_document_manager.proxy_factory" class="Sulu\Component\DocumentManager\ProxyFactory">
-            <argument type="service" id="sulu_document_manager.proxy_manager.factory.ghost"/>
-            <argument type="service" id="sulu_document_manager.event_dispatcher"/>
-            <argument type="service" id="sulu_document_manager.document_registry"/>
-            <argument type="service" id="sulu_document_manager.metadata_factory"/>
-        </service>
-
-        <service id="sulu_document_manager.proxy_manager.factory.ghost" class="ProxyManager\Factory\LazyLoadingGhostFactory"/>
-
-    </services>
-
-</container>
